Fix: Data type forced to float32 when preparing amplitude envelope data.

+++ osl_dynamics/data/processing.py
@@ -125,7 +125,8 @@ class Processing:
                         prepared_data[:, i], np.ones(n_window) / n_window, mode="same"
                     )
                     for i in range(prepared_data.shape[1])
-                ]
+                ],
+                dtype=np.float32,
             ).T
 
             # Create a memory map for the prepared data
